refactor(core): config.rs から opener マッチングを分離し apply_migrations を段階分解する
- 型: `OpenerRule` / `OpenerTool` / `OpenerPreset`
- マッチング: `find_matching_tools` / `extract_path_condition` / `extract_ext_part` /
  `split_ext_and_path`（private）/ `normalize_opener_target`（private）/
  `opener_specificity_order` / `normalize_openers`（free fn）
- プリセット検出: `detect_opener_presets` / `find_in_path`（private）/ `is_preset_already_added`

依存方向は `config.rs` → `opener.rs`（`OpenerRule`/`OpenerTool` は `Config.openers` として
config.toml に紐づく serde 型のため型定義は opener.rs に置き、config.rs が
`pub use crate::opener::{...}` で re-export）。外部呼び出し元（src-tauri の launch.rs、
snotra-settings の app.rs/tabs/opener.rs）は全て `snotra_core::config::...` パスで
参照していたため、re-export によりコード変更ゼロで動作を維持する。

逆方向として `normalize_opener_target` は `config.rs::normalize_scan_path_key` /
`normalize_extensions`（paths.scan の正規化とも共有する汎用ヘルパー）を使うため、
両関数を `pub(crate)` に昇格した。

| 旧処理（apply_migrations 内インライン） | 新 private fn |
|---|---|
| `paths.additional` → `scan` 移行の要否判定 + 実行 | `migrate_legacy_additional_paths` |
| `max_results`/`top_n_history`/`max_history_display` legacy 集約（#388） | `migrate_legacy_count_params` |
| None → Some(default) 解決 | `resolve_count_param_defaults` |
| `fuzzy_history_cap_ratio` 範囲外補正（#437） | `sanitize_fuzzy_history_cap_ratio` |
| `command` 単一文字列 → `Url` 移行 | `migrate_instant_legacy_commands` |
| system shortcut 検出時のホットキーフォールバック | `fallback_hotkey_if_system_shortcut` |

呼び出し順は元と完全に同一に固定（`|=` で全ステップを必ず評価し、`||` の
短絡評価による呼び出し漏れを回避）。真の順序依存は
`migrate_legacy_additional_paths`（scan へのエントリ追加）→
`paths.normalize_scan_paths()`（追加分を含めた dedup）の1点のみで、
コメントで明示した。他ステップは独立だが diff 最小化のため元の並びを保持。

opener 関連テスト120本中37本（find_matching_tools/normalize_opener_target/
split_ext_and_path/detect_opener_presets/is_preset_already_added 等）を
opener.rs へ移動（assertion・前提は無変更）。`Config::normalize_openers()`
メソッドをテストする3本は config.rs に残し、`make_rule` ヘルパーを
意図的に複製（テスト専用の小ヘルパーのため pub(crate) 露出を避けた）。

`find_matching_tools` のルール側正規化キャッシュ（cond.to_lowercase().replace()の
事前計算）は見送った。理由: (1) `OpenerRule`/`OpenerTool` は公開 serde 型として
テスト・snotra-settings から直接構築され正規化保証がないため、キャッシュを
効かせるには新しいラッパー型かフィールド追加が必要で複雑化する、(2) 呼び出しは
ユーザーの「開く」操作1回につき1回であり検索のようなホットパスではない
（rules 数も通常20件未満）。挙動不変・スコープ最小化を優先した。

Closes #435
