+++ snotra-core/CLAUDE.md
@@ -1,11 +1,12 @@
 # snotra-core
 
-純ロジック lib crate（12モジュール + `lib.rs`）。Win32 非依存でユニットテスト可能。
+純ロジック lib crate（13モジュール + `lib.rs`）。Win32 非依存でユニットテスト可能。
 
 ## モジュール構成
 
 - `engine.rs`: `Engine` struct（`SearchEngine` + `HistoryStore` + `Config` の facade）。`FolderListContext`（ロック外スナップショット）と `PrebuiltIndex`（インデックス高速スワップ用）を公開。**`IndexInputs`（index 構築入力 scan/show_hidden_system/show_icons/include_path_env/migemo_enabled の単一定義）と `index_stale` ledger**（`mark_index_stale` / `begin_index_drain` → snapshot / `complete_index_drain` → swap + re-diff で stale をクリア / `is_index_stale`）でコヒーレンシ判断を engine Mutex（軸1）に閉じ、config 変更→index 再構築の lost-update を塞ぐ（#347/#348-A）。`complete_index_drain` は「ビルド開始時 snapshot == 現在 IndexInputs」のときだけ stale をクリアする（ビルド中変更を取りこぼさない）
-- `config.rs`: `%APPDATA%\Snotra\config.toml` の読込/保存、既定値補完。`Language` enum（`Ja`/`En`）と `default_language()`（`sys-locale` による OS 言語自動判定、非日本語は英語フォールバック）を定義。**件数パラメータ**（#388 で役割に合わせて改名済み）: `appearance.visible_rows`=可視行数、`search.result_limit`=**検索・フォルダの結果リスト最大長**（`Engine::search`/`capture_folder_list_context` の fetch_limit）、`search.recent_limit`=空クエリ recent 件数（`recent_history`）。旧キー（`max_results`/`top_n_history`/`max_history_display`）は `apply_migrations()` が `skip_serializing` の legacy フィールド経由で後方互換移行する（2層レガシー: `result_limit` ← `[search].top_n_history` ← `[appearance].top_n_history`）。フロント `iconCacheSize` と `Config::icon_cache_cap()` はこれらから派生。実上限は名前でなく `engine.rs` の dispatch で確認する
+- `config.rs`: `%APPDATA%\Snotra\config.toml` の読込/保存、既定値補完。`Language` enum（`Ja`/`En`）と `default_language()`（`sys-locale` による OS 言語自動判定、非日本語は英語フォールバック）を定義。**件数パラメータ**（#388 で役割に合わせて改名済み）: `appearance.visible_rows`=可視行数、`search.result_limit`=**検索・フォルダの結果リスト最大長**（`Engine::search`/`capture_folder_list_context` の fetch_limit）、`search.recent_limit`=空クエリ recent 件数（`recent_history`）。旧キー（`max_results`/`top_n_history`/`max_history_display`）は `apply_migrations()` が `skip_serializing` の legacy フィールド経由で後方互換移行する（2層レガシー: `result_limit` ← `[search].top_n_history` ← `[appearance].top_n_history`）。フロント `iconCacheSize` と `Config::icon_cache_cap()` はこれらから派生。実上限は名前でなく `engine.rs` の dispatch で確認する。**`apply_migrations()` は migration 系統ごとの private fn（`migrate_legacy_additional_paths` / `migrate_legacy_count_params` / `resolve_count_param_defaults` / `sanitize_fuzzy_history_cap_ratio` / `migrate_instant_legacy_commands` / `fallback_hotkey_if_system_shortcut`）へ段階分解済み（issue #435）。呼び出し順は元と同一に固定し、`migrate_legacy_additional_paths`（`paths.additional`→`scan` 追加）→ `paths.normalize_scan_paths()`（dedup）の順序だけは真の依存（先に追加されたエントリを後続の正規化がまとめて dedup する）。他のステップは独立だが diff 最小化のため元の並びを保つ**
+- `opener.rs`: opener（外部ツール起動ルール）ターゲットの解析・正規化・マッチングエンジンと、Win 環境のプリセット検出。`OpenerRule` / `OpenerTool` / `OpenerPreset` 型、`find_matching_tools`（パス・フォルダ判定に対する最具体1ルール解決。具体度=パス条件の長さ）、`extract_path_condition` / `extract_ext_part`、`normalize_openers`（ターゲット正規化・具体度順ソート）、`opener_specificity_order`、`detect_opener_presets`（VSCode/Windows Terminal/Explorer の検出）、`is_preset_already_added` を公開。`config.rs` から分離済み（issue #435、旧 `config.rs:88-735`/`1260-1363`）。**依存方向は `config.rs` → `opener.rs`**（`OpenerRule`/`OpenerTool` は `Config.openers` として config.toml に紐づく serde 型のため型定義はこちらに置き、`config.rs` が `pub use crate::opener::{...}` で re-export して `snotra_core::config::...` の既存呼び出し元パスを維持する）。逆方向の依存として `normalize_opener_target` が `config.rs::normalize_scan_path_key` / `normalize_extensions`（`pub(crate)`、`paths.scan` の正規化とも共有する汎用ヘルパー）を使う
 - `search.rs`: 検索順位計算（Prefix/Substring/Kana/Fuzzy/Path）、履歴ブースト、incremental search キャッシュ、空クエリ時履歴候補。`SearchEngine` は並列 Vec レイアウト（`entries` / `lower_names` / `lower_file_names` / `normalized_keys` / `char_masks` / `file_name_char_masks` / `kana_lower_names`）で cache locality を確保。構築は `compute_wave1`（文字列正規化）→ `compute_wave2`（ビットマスク計算）のヘルパー関数で共通化し、`new()`（= `new_with_migemo(.., true)`）/ `new_with_migemo(entries, migemo_enabled)` / `new_with_cached_masks(.., migemo_enabled)` が共有する。**`kana_lower_names` は `migemo_enabled` が true のときのみ構築し、無効時は空 Vec**（migemo 無効ユーザーの死蔵メモリ ~2.1–2.7MB/50k を削る・構築も約 2 倍速、issue #337）。空 Vec のとき検索ループは `kana_available` 空ガードで `kana_lower_names[i]` アクセスを回避する（構築時 migemo OFF→検索時 ON の窓での panic 防止）。migemo トグルの反映は `update_config` が engine を再構築しないため、`config_watcher` が engine の `IndexInputs` 差分で `start_index_build` を kick する再構築に依存する（#347 Phase 2 で `needs_reindex` は `IndexInputs` に統合）。パスマッチング: クエリにパス区切り文字（`\` `/`）を含む場合、`normalized_key`（= `normalize_entry_key(target_path)`）に対して Substring マッチを試みる。スコアは `3000 - min(byte_pos, 500)`。name/file_name/kana 全て不成立時のフォールバック。`has_path_sep` 時は Fuzzy ビットマスク pre-filter をスキップする
 - `history.rs`: 起動履歴・クエリ別履歴・フォルダ展開履歴の管理、バイナリ永続化。**剪定容量 `top_n` は焼き込まず `save`/`save_if_dirty`/`prune` の引数で受け取る（live-read）**。`Engine` が呼び出し時に現在の config（`effective_result_limit()`）を渡すため、`result_limit` 設定変更が再起動なしで反映される（#348）。`HistoryStore` に `top_n` フィールドを再導入しないこと——焼き込むと設定変更が反映されないドリフトが復活する
 - `folder.rs`: フォルダ内列挙とフィルタ/ソート
