+++ snotra-core/src/lib.rs
@@ -6,6 +6,7 @@ pub mod folder;
 pub mod history;
 pub mod indexer;
 pub mod instant;
+pub mod opener;
 pub mod query;
 pub mod search;
 pub mod ui_types;
